% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentClass.R
\name{categorical}
\alias{categorical}
\title{Specify the items that follow a normal categorical distribution.}
\usage{
categorical(items, starting_values = "random")
}
\arguments{
\item{items}{character vector with the names of the items}

\item{starting_values}{if set to "random", starting values are generated randomly.
Alternatively, a list with matrices holding starting values for each item can be provided. Each of the
starting values matrices must have as many rows as the item has levels. Additionally,
the columns of each matrix must sum to 1.}
}
\value{
An object of class Categorical.
}
\description{
Specify the items that follow a normal categorical distribution.
}
\examples{
library(latentClass)
categorical(items = c("x1", "x2"))
# Provide starting values (assuming there are 3 classes and x1 has 2 levels, 
# while c2 has 3 levels):
categorical(items = c("x1", "x2"),
       starting_values = 
       list("x1" = matrix(c(.3, .5, .8,
                            .7, .5, .2),
                         nrow = 2,
                         byrow = TRUE,
                         dimnames = list(c("level_1", "level_2"), 
                                         c("class_1", "class_2", "class_3"))),
            "x2" = matrix(c(.1, .7, .2,
                            .8, .2, .5,
                            .1, .1, .3),
                         nrow = 3,
                         byrow = TRUE,
                         dimnames = list(c("level_1", "level_2", "level_3"),
                                         c("class_1", "class_2", "class_3"))))
)
}
