% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/composite.llike.R
\name{composite.llike}
\alias{composite.llike}
\title{Compute the composite likelihood}
\usage{
composite.llike(dat, param, n.class = 2, margin.size = 3)
}
\arguments{
\item{dat}{The data array of the format \code{(r[1],...,r[m])}}

\item{param}{The vector of parameters of the model. It represents elements of each transition matrix apart from the last column.}

\item{n.class}{The number of hidden classes}

\item{margin.size}{The size of the margins that are used to compute the composite lieklihood.}
}
\description{
This internal function computes the composite likelihood for a given latent class model
and a collection of margins.
}

