% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/constraints.theta.R
\name{constraints.theta}
\alias{constraints.theta}
\title{constraints.theta function}
\usage{
constraints.theta(r, n.class)
}
\arguments{
\item{r}{The vector of dimensions of the format \code{(r[1],...,r[m])}}

\item{n.class}{The number of latent classes}
}
\description{
This auxiliary function outputs for any latent class model the corresponding matrix that encodes paramameter constraints.
The output is a list \code{A,b}, where \code{A} is a matrix and \code{b} is a vector such that \code{A\%*\% theta >= b} is the set of the inequalities defining parameter constraints.
}
\details{
In addition, for identifiability issues, we require that teh latent probabilities are sorted
from the largest to the smallest.
}
\examples{
constraints.theta(c(2,2,2,2),2)
}

