% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/E.step.R
\name{E.step}
\alias{E.step}
\title{E-step of the EM algorithm}
\usage{
E.step(phat, theta)
}
\arguments{
\item{phat}{The array of normalized counts (sample proportions) of format (r1,...,rm)}

\item{theta}{The parameter vector in form of a list of size m+1. The last element is the marginal distribution of the latent variable. The other elements of the list correspond to conditional probabilities of the observed variables given the latent one.}
}
\description{
This internal function performs an E-step of the EM-algorithm
}
\examples{
E.step(phat,theta)
}
\keyword{E-step}

