% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/exact222.R
\name{exact222}
\alias{exact222}
\title{Exact maximum likelihood for three binary variables}
\usage{
exact222(counts, tol = 1e-08)
}
\arguments{
\item{counts}{A (2,2,2) array with counts.
tol   tolerance with respect to which model membership is checked (default tol=1e-8)}
}
\description{
This function computes the exact maximum likelihood for the model with three observed binary variables and two latent classes.
}
\examples{
n <-  c(12, 23, 34, 45, 56, 67, 78, 90)
res <- exact222(n)
res$min.deviance
res$fitted.values
}
\keyword{MLE}
\keyword{exact}

