\name{latentClass-package}
\alias{latentClass-package}
\alias{latentClass}
\docType{package}
\title{
\packageTitle{latentClass}
}
\description{
\packageDescription{latentClass}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{latentClass}

Maintainer: \packageMaintainer{latentClass}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = latentClass)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{latentClass}}
%\section{Documentation Index}{\packageIndices{latentClass}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
