% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/LC.EM.R
\name{LC.EM}
\alias{LC.EM}
\title{LC.EM function}
\usage{
LC.EM(counts, k, tries = 3, theta = NULL, tol = 1e-06)
}
\arguments{
\item{counts}{The array of counts of format (r[1],...,r[m])}

\item{k}{The number of latent classes fitted}

\item{tries}{The number of times the EM algorithm reruns from different random starting points. The default value is tries=3}

\item{theta}{The vector of parameters from which the algorithm starts. If not specified, the algorithm starts from a random point.}

\item{tol}{The convergence criterion for the EM algorithm. The maximal decrease of the log-likelihood function that will terminate the algorithm.}
}
\description{
This function estimates parameters of the latent class model using the standard EM algorithm.
}
\examples{
theta0 <- list()
length(theta0) <- 5
theta0[[1]] <- matrix(c(0.8,1-0.7,1-0.8,0.7),2,2)
theta0[[2]] <- matrix(c(0.8,1-0.7,1-0.8,0.7),2,2)
theta0[[3]] <- matrix(c(0.8,1-0.7,1-0.8,0.7),2,2)
theta0[[4]] <- matrix(c(0.8,1-0.9,1-0.8,0.9),2,2)
theta0[[5]] <- c(1-0.7,0.7)
n <- 1000
counts <- sample.counts(n, theta0)
EM(counts,2)
}
\keyword{EM}
\keyword{algorithm,}
\keyword{class}
\keyword{latent}
\keyword{model}

