% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/llike.R
\name{llike}
\alias{llike}
\title{The log-likelihood function}
\usage{
llike(counts, theta = NULL)
}
\arguments{
\item{counts}{The array of counts of format (r[1],...,r[m])}

\item{theta}{The vector of parameters in form of a list of size m+1, where the the first m entries are kxr[i] matrices of teh conditional distribution of an observed variable given the latent one. The last element is the vector of the marginal distribution of the latent variable. If not provided then the log-likelihood function at the sample distribution is computed.}
}
\description{
This function computes the value of the observed log-likelihood function either for a given parameter theta or for the sample distribution.
}
\examples{
llike(counts,theta=NULL)
}
\keyword{distribution}
\keyword{function,}
\keyword{log-likelihood}
\keyword{sample}

