% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/M.step.R
\name{M.step}
\alias{M.step}
\title{M-step of the EM algorithm}
\usage{
M.step(phat, fphat)
}
\arguments{
\item{phat}{The array of normalized counts (sample proportions) of format (r[1],...,r[m])}

\item{fphat}{The list of k arrays that represent the full distribution over (r[1],...,r[m],k)}
}
\description{
This internal function performs the M-step of the EM-algorithm
}
\examples{
M.step(phat,theta)
}
\keyword{E-step}

