% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentClass.R
\name{normal}
\alias{normal}
\title{Specify the items that follow a normal (Gaussian) distribution.}
\usage{
normal(items, sd_equal = TRUE, starting_values = "random")
}
\arguments{
\item{items}{character vector with the names of the items}

\item{sd_equal}{if set to TRUE, the standard deviations for each item
will be set to equal across classes. Alternatively, a vector can be passed 
that specifies for each of the items if the standard deviations should be equal
across classes}

\item{starting_values}{if set to "random", starting values are generated randomly.
Alternatively, a list with matrices holding starting values for each item can be provided. Each of the
starting values matrices must have 2 rows (starting values for mean and standard deviation)
and as many columns as there are classes}
}
\value{
An object of class Normal.
}
\description{
For all items specified here, latentClass will use a normal distribution
density as likelihood.
}
\examples{
library(latentClass)
normal(items = c("x1", "x2"))
# Allow for class-specific standard deviations:
normal(items = c("x1", "x2"),
       sd_equal = FALSE)
# Allow for class-specific standard deviations for x1, but not x2:
normal(items = c("x1", "x2"),
       sd_equal = c(TRUE, FALSE))
# Provide starting values (assuming there are 3 classes):
normal(items = c("x1", "x2"),
       starting_values = 
       list("x1" = matrix(c(-3, 0, 3,
                             1, 1, 1),
                         nrow = 2,
                         byrow = TRUE,
                         dimnames = list(c("mean", "sd"), 
                                         c("class_1", "class_2", "class_3"))),
            "x2" = matrix(c(-1, 0, 1,
                             1, 1, 1),
                         nrow = 2,
                         byrow = TRUE,
                         dimnames = list(c("mean", "sd"), 
                                         c("class_1", "class_2", "class_3")))))
}
