% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentClass.R
\name{optimizer_settings}
\alias{optimizer_settings}
\title{Fine tune the optimizer settings.}
\usage{
optimizer_settings(max_iter = 1000, convergence_criterion = 1e-10)
}
\arguments{
\item{max_iter}{maximum number of iterations}

\item{convergence_criterion}{cut-off value. If the relative change in the
log-likelihood from one iteration to the next falls below this value, the optimization
stops and the model is seen as converged.}
}
\value{
list with optimizer settings
}
\description{
Fine tune the optimizer settings.
}
\examples{
library(latentClass)
optimizer_settings()
}
