% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentClass.R
\name{print.latentClass}
\alias{print.latentClass}
\title{print.latentClass}
\usage{
\method{print}{latentClass}(x, ...)
}
\arguments{
\item{x}{A latent class model object.}

\item{...}{not used}
}
\value{
nothing
}
\description{
Print the results of a latent class model.
}
