% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sample.Theta.R
\name{sample.Theta}
\alias{sample.Theta}
\title{Sample a parameter vector}
\usage{
sample.Theta(r, k)
}
\arguments{
\item{r}{A vector  (r[1],...,r[m]) of integers, where m is the number of observed variables and r[i] is teh statespace of the i-th variable.}

\item{k}{The number of latent classes fitted}
}
\description{
This function samples a parameter vector for a latent class model. Each row from each conditional distribution matrix is sampled from the uniform Dirichlet distribution.
}
\examples{
r <- c(2,2,2,2)
k <- 3
sample.Theta(r,k)
}
\keyword{parameters}
\keyword{sample}

