% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_latent_class_data}
\alias{simulate_latent_class_data}
\title{simulate_latent_class_data}
\usage{
simulate_latent_class_data()
}
\value{
data frame with cat_1 and cat_2 (continuous variables) and norm_1 and norm_2 (continuous variables).
}
\description{
Generates simulated data for a latent class model.
}
