% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentClass.R
\name{summary.latentClass}
\alias{summary.latentClass}
\title{summary.latentClass}
\usage{
\method{summary}{latentClass}(object, ...)
}
\arguments{
\item{object}{A latent class model object.}

\item{...}{not used}
}
\value{
nothing
}
\description{
Print the results of a latent class model.
}
