% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/theta.list.R
\name{theta.list}
\alias{theta.list}
\title{Convert the parameter vector to a list in the internal format}
\usage{
theta.list(param, r, n.class = 2)
}
\arguments{
\item{param}{The vector of parameters of the model. It represents elements of each transition matrix apart from the last column.}

\item{r}{The dimension vector of the data (r[1],...,r[m])}

\item{n.class}{The number of hidden classes}
}
\description{
This internal function converts a parameter vector to a prefered list format
}
\examples{
theta.list(param,r,class=2)
}
\keyword{theta}

