% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/theta.unlist.R
\name{theta.unlist}
\alias{theta.unlist}
\title{M-step of the EM algorithm}
\usage{
theta.unlist(theta)
}
\arguments{
\item{phat}{The array of normalized counts (sample proportions) of format (r[1],...,r[m])}

\item{fphat}{The list of k arrays that represent the full distribution over (r[1],...,r[m],k)}
}
\description{
This internal function converts a parameter vector to a prefered list format
}
\examples{
theta.list(param,r,class=2)
}
\keyword{E-step}

